# services/catalog-service/src/main/resources/static/openapi/catalog-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mundiapolis Catalog API",
    "version": "1.5.0",
    "description": "Authoritative catalog metadata and privacy-safe review commands and reads. Copy counts and returned-loan eligibility are versioned projections owned by Circulation."
  },
  "paths": {
    "/api/v1/catalog/works": {
      "post": {
        "operationId": "createWork",
        "x-required-scopes": ["catalog.manage"],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateWorkRequest" } } } },
        "responses": {
          "201": { "description": "Work created or exact command replayed", "headers": { "Location": { "description": "Canonical work resource.", "schema": { "type": "string", "format": "uri-reference" } }, "ETag": { "description": "Quoted aggregate version.", "schema": { "type": "string" } }, "Idempotency-Replayed": { "schema": { "type": "boolean" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogCommandResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" }
        }
      }
    },
    "/api/v1/catalog/works/{workId}/editions": {
      "post": {
        "operationId": "createEdition",
        "x-required-scopes": ["catalog.manage"],
        "parameters": [
          { "$ref": "#/components/parameters/WorkId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateEditionRequest" } } } },
        "responses": {
          "201": { "description": "Edition created or exact command replayed", "headers": { "Location": { "description": "Canonical edition resource.", "schema": { "type": "string", "format": "uri-reference" } }, "ETag": { "description": "Quoted aggregate version.", "schema": { "type": "string" } }, "Idempotency-Replayed": { "schema": { "type": "boolean" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogCommandResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" }
        }
      }
    },
    "/api/v1/catalog/works/{workId}": {
      "get": {
        "operationId": "getWork",
        "x-required-scopes": ["catalog.read"],
        "parameters": [{ "$ref": "#/components/parameters/WorkId" }],
        "responses": {
          "200": { "description": "Work found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Work" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "operationId": "updateWork",
        "x-required-scopes": ["catalog.manage"],
        "parameters": [
          { "$ref": "#/components/parameters/WorkId" },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateWorkRequest" } } } },
        "responses": {
          "200": { "description": "Work updated or exact command replayed", "headers": { "ETag": { "description": "Quoted aggregate version.", "schema": { "type": "string" } }, "Idempotency-Replayed": { "schema": { "type": "boolean" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogCommandResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" }
        }
      }
    },
    "/api/v1/catalog/editions/{editionId}": {
      "get": {
        "operationId": "getEdition",
        "x-required-scopes": ["catalog.read"],
        "parameters": [{ "$ref": "#/components/parameters/EditionId" }],
        "responses": {
          "200": { "description": "Edition found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Edition" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "operationId": "updateEdition",
        "x-required-scopes": ["catalog.manage"],
        "parameters": [
          { "$ref": "#/components/parameters/EditionId" },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateEditionRequest" } } } },
        "responses": {
          "200": { "description": "Edition updated or exact command replayed", "headers": { "ETag": { "description": "Quoted aggregate version.", "schema": { "type": "string" } }, "Idempotency-Replayed": { "schema": { "type": "boolean" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogCommandResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" }
        }
      }
    },
    "/api/v1/catalog/works/{workId}/reviews": {
      "get": {
        "operationId": "getPublishedReviews",
        "x-required-scopes": ["catalog.read"],
        "parameters": [
          { "$ref": "#/components/parameters/WorkId" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 0, "maximum": 10000, "default": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Published review page without member identifiers", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogReviewPage" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "operationId": "createReview",
        "x-required-scopes": ["catalog.review.write"],
        "parameters": [
          { "$ref": "#/components/parameters/WorkId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateReviewRequest" } } } },
        "responses": {
          "201": { "description": "Review created or exact command replayed", "headers": { "Location": { "description": "Canonical review command resource.", "schema": { "type": "string", "format": "uri-reference" } }, "ETag": { "description": "Quoted aggregate version.", "schema": { "type": "string" } }, "Idempotency-Replayed": { "schema": { "type": "boolean" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogCommandResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
    "/api/v1/catalog/reviews/{reviewId}": {
      "put": {
        "operationId": "updateReview",
        "x-required-scopes": ["catalog.review.write"],
        "parameters": [
          { "$ref": "#/components/parameters/ReviewId" },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateReviewRequest" } } } },
        "responses": {
          "200": { "description": "Review updated or exact command replayed", "headers": { "ETag": { "description": "Quoted aggregate version.", "schema": { "type": "string" } }, "Idempotency-Replayed": { "schema": { "type": "boolean" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogCommandResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" }
        }
      },
      "delete": {
        "operationId": "deleteReview",
        "x-required-scopes": ["catalog.review.write"],
        "parameters": [
          { "$ref": "#/components/parameters/ReviewId" },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "responses": {
          "200": { "description": "Review deleted or exact command replayed", "headers": { "ETag": { "description": "Quoted terminal aggregate version.", "schema": { "type": "string" } }, "Idempotency-Replayed": { "schema": { "type": "boolean" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogCommandResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/catalog/editions/{editionId}/activation": {
      "post": {
        "operationId": "setEditionActive",
        "x-required-scopes": ["catalog.manage"],
        "parameters": [
          { "$ref": "#/components/parameters/EditionId" },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SetEditionActiveRequest" } } } },
        "responses": {
          "200": { "description": "Edition activation changed or exact command replayed", "headers": { "ETag": { "description": "Quoted aggregate version.", "schema": { "type": "string" } }, "Idempotency-Replayed": { "schema": { "type": "boolean" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogCommandResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" }
        }
      }
    },
...
    "/api/v1/catalog/learning-resource-imports/{importId}": {
      "get": {
        "operationId": "getLearningResourceImportEvidence",
        "x-required-scopes": ["catalog.learning-resource.import"],
        "parameters": [{ "name": "importId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Completed immutable import evidence", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LearningResourceImportResult" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "operationId": "importLearningResources",
        "x-required-scopes": ["catalog.learning-resource.import"],
        "parameters": [{ "name": "importId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LearningResourceImportRequest" } } } },
        "responses": {
          "200": { "description": "Atomic import completed or exactly replayed", "headers": { "Idempotency-Replayed": { "schema": { "type": "boolean" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LearningResourceImportResult" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "WorkId": { "name": "workId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "EditionId": { "name": "editionId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 16, "maxLength": 128, "pattern": "^[ -~]+$" } },
      "IfMatch": { "name": "If-Match", "in": "header", "required": true, "description": "Exact quoted aggregate version, for example \"0\".", "schema": { "type": "string", "pattern": "^\"(0|[1-9][0-9]*)\"$" } }
    },
    "responses": {
      "BadRequest": { "description": "Invalid request" },
      "Unauthorized": { "description": "Missing or invalid bearer token" },
      "Forbidden": { "description": "Required scope is missing" },
      "NotFound": { "description": "Resource not found" },
      "Conflict": { "description": "Aggregate, contributor metadata, or idempotency conflict" },
      "PayloadTooLarge": { "description": "Command body exceeds 128 KiB" },
...
    },
    "schemas": {
...
      "Work": {
        "type": "object",
        "additionalProperties": false,
        "required": ["workId", "title", "summary", "description", "genre", "rating", "authors"],
        "properties": {
          "workId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "summary": { "type": "string" },
          "description": { "type": "string" },
          "genre": { "type": "string" },
          "rating": { "type": "number", "minimum": 0, "maximum": 5 },
          "authors": { "type": "array", "items": { "$ref": "#/components/schemas/Author" } }
        }
      },
...
      "CatalogReviewPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["reviews", "total", "page", "totalPages"],
        "properties": {
          "reviews": { "type": "array", "items": { "$ref": "#/components/schemas/CatalogReview" } },
          "total": { "type": "integer", "minimum": 0 },
          "page": { "type": "integer", "minimum": 0 },
          "totalPages": { "type": "integer", "minimum": 0 }
        }
      },
...
      "CreateWorkRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["workId", "title", "summary", "description", "genre", "authors", "reason"],
        "properties": {
          "workId": { "type": "string", "format": "uuid" },
          "title": { "type": "string", "maxLength": 500 },
          "summary": { "type": "string", "maxLength": 1000 },
          "description": { "type": "string", "maxLength": 10000 },
          "genre": { "type": "string", "maxLength": 120 },
          "authors": { "type": "array", "minItems": 1, "maxItems": 20, "items": { "$ref": "#/components/schemas/CreateAuthorRequest" } },
          "reason": { "type": "string", "minLength": 8, "maxLength": 500 }
        }
      },
      "CreateEditionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["editionId", "title", "isbn", "publisher", "publicationYear", "language", "pageCount", "isActive", "reason"],
        "properties": {
          "editionId": { "type": "string", "format": "uuid" },
          "title": { "type": "string", "maxLength": 500 },
          "isbn": { "type": "string", "maxLength": 32 },
          "publisher": { "type": "string", "maxLength": 300 },
          "publicationYear": { "type": "integer", "minimum": 1000, "maximum": 3000 },
          "language": { "type": "string", "maxLength": 80 },
          "pageCount": { "type": "integer", "minimum": 1, "maximum": 100000 },
          "coverUrl": { "type": ["string", "null"], "format": "uri", "maxLength": 2048 },
          "coverColor": { "type": ["string", "null"], "pattern": "^#[0-9A-Fa-f]{6}$" },
          "videoUrl": { "type": ["string", "null"], "format": "uri", "maxLength": 2048 },
          "isActive": { "type": "boolean", "default": true },
          "reason": { "type": "string", "minLength": 8, "maxLength": 500 }
        }
      },
      "UpdateWorkRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["title", "summary", "description", "genre", "authors", "reason"],
        "properties": {
          "title": { "type": "string", "maxLength": 500 },
          "summary": { "type": "string", "maxLength": 1000 },
          "description": { "type": "string", "maxLength": 10000 },
          "genre": { "type": "string", "maxLength": 120 },
          "authors": { "type": "array", "minItems": 1, "maxItems": 20, "items": { "$ref": "#/components/schemas/CreateAuthorRequest" } },
          "reason": { "type": "string", "minLength": 8, "maxLength": 500 }
        }
      },
...
      "CreateReviewRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["rating", "content"],
        "properties": {
          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "content": { "type": "string", "minLength": 1, "maxLength": 4000 }
        }
      },
...
      "CatalogCommandResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["aggregateType", "aggregateId", "aggregateVersion", "occurredAt"],
        "properties": {
          "aggregateType": { "type": "string", "enum": ["work", "edition", "review"] },
          "aggregateId": { "type": "string", "format": "uuid" },
          "aggregateVersion": { "type": "integer", "minimum": 0 },
          "occurredAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }]
}